fix(responses): emit created_at as integer Unix seconds (v0.7.2)
The Responses API documents created_at as an integer, but the converter
emitted a float (e.g. 1782085884.1576056). Downstream consumers that decode
the field into a strict int — notably new-api's Go struct
OpenAIResponsesResponse.created_at — rejected the payload with
"json: cannot unmarshal number … into Go struct field … of type int",
causing channel-test failures for glm-5.2 and deepseek-v4-pro on the
/v1/responses endpoint. _now() now returns int(time.time()) and the
non-streaming + streaming Responses paths both emit integers.

+++ src/local_openai2anthropic/responses_web_search.py
@@ -341,7 +341,7 @@ async def _stream_responses_with_web_search(
     from local_openai2anthropic.responses_converter import _sse
 
     rid = responses_payload.get("id", "")
-    created_at = responses_payload.get("created_at", float(time.time()))
+    created_at = responses_payload.get("created_at", int(time.time()))
     status = responses_payload.get("status", "completed")
 
     # response.created / in_progress
